requirements.txt
langchain_openai==0.3.18
python-dotenv==1.1.0
pydantic==2.11.5
langgraph==0.4.7
langchain_community==0.3.24
langchain-mcp-adapters==0.1.1
jinja2==2-3.1.6
markdownify==1.1.0
readabilipy==0.3.0
langchain_experimental==0.3.4